+++ tests/testthat/test-time.R
@@ -106,12 +106,12 @@ test_with_dir("runtime predictions", {
   known_times <- c(
     a = 0, b = 0, c = 0, f = 0, g = 0, h = 0, i = 0, j = 0,
     readRDS = 0, saveRDS = 0,
-    "\"saveRDS\"" = 0, "\"input.rds\"" = 0,
     myinput = 10,
     nextone = 33,
     yourinput = 27,
     final = Inf
   )
+  known_times[reencode_path(c("saveRDS", "input.rds"))] <- 0
   targets <- c("nextone", "yourinput")
   p5 <- predict_runtime(
     config = con,